requirements.txt
streamlit==1.44.1
google-generativeai==0.8.5
requests==2.32.3
beautifulsoup4==4.13.4
python-dotenv==1.1.0
gitpython==3.1.44
altair==5.5.0
pandas==2.2.3
numpy==2.2.5
